@@ -1,15 +1,15 @@
 import { TestBed, async } from '@angular/core/testing';
-
 import { AppComponent } from './app.component';
 
 describe('AppComponent', () => {
-  beforeEach(async(() => {
+  beforeEach(() => {
     TestBed.configureTestingModule({
       declarations: [
         AppComponent
       ],
-    }).compileComponents();
-  }));
+    });
+    TestBed.compileComponents();
+  });
 
   it('should create the app', async(() => {
     const fixture = TestBed.createComponent(AppComponent);
